+++ x/vesting/types/clawback_vesting_account.go
@@ -24,12 +24,15 @@ func NewClawbackVestingAccount(
 	lockupPeriods,
 	vestingPeriods sdkvesting.Periods,
 ) *ClawbackVestingAccount {
-	// copy and align schedules to avoid mutating inputs
+	// copy and align schedules to the same start time to
+	// avoid mutating inputs
 	lp := make(sdkvesting.Periods, len(lockupPeriods))
 	copy(lp, lockupPeriods)
 	vp := make(sdkvesting.Periods, len(vestingPeriods))
 	copy(vp, vestingPeriods)
+
 	_, endTime := AlignSchedules(startTime.Unix(), startTime.Unix(), lp, vp)
+
 	baseVestingAcc := &sdkvesting.BaseVestingAccount{
 		BaseAccount:     baseAcc,
 		OriginalVesting: originalVesting,
@@ -186,12 +189,14 @@ func (va ClawbackVestingAccount) ComputeClawback(
 	// Cap the unlocking schedule to the new total vested.
 	//  - If lockup has already passed, all vested coins are unlocked.
 	//  - If lockup has not passed, the vested coins, are still locked.
-	capPeriods := []sdkvesting.Period{
+	capPeriods := sdkvesting.Periods{
 		{
 			Length: 0,
 			Amount: totalVested,
 		},
 	}
+
+	// minimum of the 2 periods
 	_, newLockingEnd, newLockupPeriods := ConjunctPeriods(va.GetStartTime(), va.GetStartTime(), va.LockupPeriods, capPeriods)
 
 	// Now construct the new account state