+++ x/vesting/types/schedule.go
@@ -14,27 +14,30 @@ import (
 // ReadSchedule returns the value of a schedule at readTime.
 func ReadSchedule(
 	startTime, endTime int64,
-	periods []sdkvesting.Period,
+	periods sdkvesting.Periods,
 	totalCoins sdk.Coins,
 	readTime int64,
 ) sdk.Coins {
+	// return empty coins if the read time is before or equal start time
 	if readTime <= startTime {
 		return sdk.NewCoins()
 	}
+	// return the total coins when the read time is equal or after
+	// end time
 	if readTime >= endTime {
 		return totalCoins
 	}
 
-	coins := sdk.NewCoins() // sum of amounts for events before readTime
-	time := startTime
+	coins := sdk.Coins{} // sum of amounts for events before readTime
+	elapsedTime := startTime
 
 	for _, period := range periods {
-		if readTime < time+period.Length {
+		if readTime < elapsedTime+period.Length {
 			// we're reading before the next event
 			break
 		}
 		coins = coins.Add(period.Amount...)
-		time += period.Length
+		elapsedTime += period.Length
 	}
 
 	return coins
@@ -43,26 +46,33 @@ func ReadSchedule(
 // ReadPastPeriodCount returns the amount of passed periods before read time
 func ReadPastPeriodCount(
 	startTime, endTime int64,
-	periods []sdkvesting.Period,
+	periods sdkvesting.Periods,
 	readTime int64,
 ) int {
 	passedPeriods := 0
+
+	// return 0 if the read time is before or equal start time
 	if readTime <= startTime {
-		return passedPeriods
+		return 0
 	}
+
+	// return all the periods when the read time is equal or after
+	// end time
 	if readTime >= endTime {
 		return len(periods)
 	}
 
-	time := startTime
+	elapsedTime := startTime
 
+	// for every period, add the period length to the elapsed time until
+	// the read time is before the next period
 	for _, period := range periods {
-		if readTime < time+period.Length {
+		if readTime < elapsedTime+period.Length {
 			// we're reading before the next event
 			break
 		}
 		passedPeriods++
-		time += period.Length
+		elapsedTime += period.Length
 	}
 
 	return passedPeriods
@@ -74,98 +84,105 @@ func ReadPastPeriodCount(
 // their start times and periods. Returns new start time, new end time, and
 // merged vesting events, relative to the new start time.
 func DisjunctPeriods(
-	startP, startQ int64,
-	periodsP, periodsQ []sdkvesting.Period,
-) (int64, int64, []sdkvesting.Period) {
-	timeP := startP // time of last merged p event, next p event is relative to this time
-	timeQ := startQ // time of last merged q event, next q event is relative to this time
-	iP := 0         // p indexes before this have been merged
-	iQ := 0         // q indexes before this have been merged
-	lenP := len(periodsP)
-	lenQ := len(periodsQ)
-	startTime := Min64(startP, startQ) // we pick the earlier time
-	time := startTime                  // time of last merged event, or the start time
-	merged := []sdkvesting.Period{}
+	startTimePeriodsA, startTimePeriodsB int64,
+	periodsA, periodsB sdkvesting.Periods,
+) (startTime, endTime int64, periods sdkvesting.Periods) {
+	timePeriodA := startTimePeriodsA  // time of last merged periods A event, next p event is relative to this time
+	timePeriodsB := startTimePeriodsB // time of last merged periods B event, next periodsB event is relative to this time
+
+	idxPeriodsA := 0 // periods A indexes before this have been merged
+	idxPeriodsB := 0 // periods B indexes before this have been merged
+
+	lenPeriodsA := len(periodsA)
+	lenPeriodsB := len(periodsB)
+	startTime = Min64(startTimePeriodsA, startTimePeriodsB) // we pick the earlier time
+	endTime = startTime                                     // time of last merged event, or the start time
+	periods = sdkvesting.Periods{}                          // merged periods
 
 	// emit adds an output period and updates the last event time
 	emit := func(nextTime int64, amount sdk.Coins) {
 		period := sdkvesting.Period{
-			Length: nextTime - time,
+			Length: nextTime - endTime,
 			Amount: amount,
 		}
-		merged = append(merged, period)
-		time = nextTime
+		periods = append(periods, period)
+		endTime = nextTime
 	}
 
-	// consumeP emits the next period from p, updating indexes
-	consumeP := func(nextP int64) {
-		emit(nextP, periodsP[iP].Amount)
-		timeP = nextP
-		iP++
+	// consumeA emits the next period from A, updating indexes
+	consumeA := func(nextPeriodA int64) {
+		emit(nextPeriodA, periodsA[idxPeriodsA].Amount)
+		timePeriodA = nextPeriodA
+		idxPeriodsA++
 	}
 
-	// consumeQ emits the next period from q, updating indexes
-	consumeQ := func(nextQ int64) {
-		emit(nextQ, periodsQ[iQ].Amount)
-		timeQ = nextQ
-		iQ++
+	// consumeB emits the next period from B, updating indexes
+	consumeB := func(nextPeriodB int64) {
+		emit(nextPeriodB, periodsB[idxPeriodsB].Amount)
+		timePeriodsB = nextPeriodB
+		idxPeriodsB++
 	}
 
-	// consumeBoth emits a merge of the next periods from p and q, updating indexes
+	// consumeBoth emits a merge of the next periods from periods A and B, updating indexes
 	consumeBoth := func(nextTime int64) {
-		emit(nextTime, periodsP[iP].Amount.Add(periodsQ[iQ].Amount...))
-		timeP = nextTime
-		timeQ = nextTime
-		iP++
-		iQ++
+		emit(nextTime, periodsA[idxPeriodsA].Amount.Add(periodsB[idxPeriodsB].Amount...))
+		timePeriodA = nextTime
+		timePeriodsB = nextTime
+		idxPeriodsA++
+		idxPeriodsB++
 	}
 
 	// while there are more events in both schedules, handle the next one, merge
 	// if concurrent
-	for iP < lenP && iQ < lenQ {
-		nextP := timeP + periodsP[iP].Length // next p event in absolute time
-		nextQ := timeQ + periodsQ[iQ].Length // next q event in absolute time
+	for idxPeriodsA < lenPeriodsA && idxPeriodsB < lenPeriodsB {
+		nextPeriodA := timePeriodA + periodsA[idxPeriodsA].Length  // next periodsA event in absolute time
+		nextPeriodB := timePeriodsB + periodsB[idxPeriodsB].Length // next periodsB event in absolute time
 		switch {
-		case nextP < nextQ:
-			consumeP(nextP)
-		case nextP > nextQ:
-			consumeQ(nextQ)
+		case nextPeriodA < nextPeriodB:
+			consumeA(nextPeriodA)
+		case nextPeriodA > nextPeriodB:
+			consumeB(nextPeriodB)
 		default:
-			consumeBoth(nextP)
+			consumeBoth(nextPeriodA)
 		}
 	}
-	// consume remaining events in schedule P
-	for iP < lenP {
-		nextP := timeP + periodsP[iP].Length
-		consumeP(nextP)
+	// consume remaining events in schedule Periods A
+	for idxPeriodsA < lenPeriodsA {
+		nextPeriodA := timePeriodA + periodsA[idxPeriodsA].Length
+		consumeA(nextPeriodA)
 	}
-	// consume remaining events in schedule Q
-	for iQ < lenQ {
-		nextQ := timeQ + periodsQ[iQ].Length
-		consumeQ(nextQ)
+	// consume remaining events in schedule PeriodsB
+	for idxPeriodsB < lenPeriodsB {
+		nextPeriodB := timePeriodsB + periodsB[idxPeriodsB].Length
+		consumeB(nextPeriodB)
 	}
-	return startTime, time, merged
+
+	return startTime, endTime, periods
 }
 
 // ConjunctPeriods returns the combination of two period schedules where the
 // result is the minimum of the two schedules.
+// It returns the resulting periods start and end times as well as the resulting
+// conjunction periods.
+// TODO: rename and add comprehensive comments, this is currently not maintainable
 func ConjunctPeriods(
-	startP, startQ int64,
-	periodsP,
-	periodsQ []sdkvesting.Period,
-) (startTime int64, endTime int64, merged []sdkvesting.Period) {
-	timeP := startP
-	timeQ := startQ
-	iP := 0
-	iQ := 0
-	lenP := len(periodsP)
-	lenQ := len(periodsQ)
-	startTime = Min64(startP, startQ)
+	startTimePeriodA, startTimePeriodB int64,
+	periodsA, periodsB sdkvesting.Periods,
+) (startTime, endTime int64, conjunctionPeriods sdkvesting.Periods) {
+	timePeriodsA := startTimePeriodA
+	timePeriodsB := startTimePeriodB
+	idxPeriodsA := 0
+	idxPeriodsB := 0
+	lenPeriodsA := len(periodsA)
+	lenPeriodsB := len(periodsB)
+	startTime = Min64(startTimePeriodA, startTimePeriodB)
 	time := startTime
-	merged = []sdkvesting.Period{}
-	amount := sdk.NewCoins()
-	amountP := amount
-	amountQ := amount
+
+	conjunctionPeriods = sdkvesting.Periods{}
+	amount := sdk.Coins{}
+
+	totalAmountPeriodsA := amount
+	totalAmountPeriodsB := amount
 
 	// emit adds an output period and updates the last event time
 	emit := func(nextTime int64, coins sdk.Coins) {
@@ -173,113 +190,120 @@ func ConjunctPeriods(
 			Length: nextTime - time,
 			Amount: coins,
 		}
-		merged = append(merged, period)
+		conjunctionPeriods = append(conjunctionPeriods, period)
 		time = nextTime
 		amount = amount.Add(coins...)
 	}
 
-	// consumeP processes the next event in P and emits an event
+	// consumeA processes the next event in P and emits an event
 	// if the minimum of P and Q changes
-	consumeP := func(nextTime int64) {
-		amountP = amountP.Add(periodsP[iP].Amount...)
-		min := amountP.Min(amountQ)
+	consumeA := func(nextTime int64) {
+		totalAmountPeriodsA = totalAmountPeriodsA.Add(periodsA[idxPeriodsA].Amount...)
+		min := totalAmountPeriodsA.Min(totalAmountPeriodsB)
 		if amount.IsAllLTE(min) {
 			diff := min.Sub(amount)
 			if !diff.IsZero() {
 				emit(nextTime, diff)
 			}
 		}
-		timeP = nextTime
-		iP++
+		timePeriodsA = nextTime
+		idxPeriodsA++
 	}
 
-	// consumeQ processes the next event in Q and emits an event
+	// consumeB processes the next event in Q and emits an event
 	// if the minimum of P and Q changes
-	consumeQ := func(nextTime int64) {
-		amountQ = amountQ.Add(periodsQ[iQ].Amount...)
-		min := amountP.Min(amountQ)
+	consumeB := func(nextTime int64) {
+		totalAmountPeriodsB = totalAmountPeriodsB.Add(periodsB[idxPeriodsB].Amount...)
+		min := totalAmountPeriodsA.Min(totalAmountPeriodsB)
 		if amount.IsAllLTE(min) {
 			diff := min.Sub(amount)
 			if !diff.IsZero() {
 				emit(nextTime, diff)
 			}
 		}
-		timeQ = nextTime
-		iQ++
+		timePeriodsB = nextTime
+		idxPeriodsB++
 	}
 
 	// consumeBoth processes simultaneous events in P and Q and emits an
 	// event if the minimum of P and Q changes
 	consumeBoth := func(nextTime int64) {
-		amountP = amountP.Add(periodsP[iP].Amount...)
-		amountQ = amountQ.Add(periodsQ[iQ].Amount...)
-		min := amountP.Min(amountQ)
+		totalAmountPeriodsA = totalAmountPeriodsA.Add(periodsA[idxPeriodsA].Amount...)
+		totalAmountPeriodsB = totalAmountPeriodsB.Add(periodsB[idxPeriodsB].Amount...)
+		min := totalAmountPeriodsA.Min(totalAmountPeriodsB)
 		if amount.IsAllLTE(min) {
 			diff := min.Sub(amount)
 			if !diff.IsZero() {
 				emit(nextTime, diff)
 			}
 		}
-		timeP = nextTime
-		timeQ = nextTime
-		iP++
-		iQ++
+		timePeriodsA = nextTime
+		timePeriodsB = nextTime
+		idxPeriodsA++
+		idxPeriodsB++
 	}
 
 	// while there are events left in both schedules, process the next one
-	for iP < lenP && iQ < lenQ {
-		nextP := timeP + periodsP[iP].Length // next p event in absolute time
-		nextQ := timeQ + periodsQ[iQ].Length // next q event in absolute time
+	for idxPeriodsA < lenPeriodsA && idxPeriodsB < lenPeriodsB {
+		nextPeriodA := timePeriodsA + periodsA[idxPeriodsA].Length // next periods A event in absolute time
+		nextPeriodB := timePeriodsB + periodsB[idxPeriodsB].Length // next periods B event in absolute time
 		switch {
-		case nextP < nextQ:
-			consumeP(nextP)
-		case nextP > nextQ:
-			consumeQ(nextQ)
+		case nextPeriodA < nextPeriodB:
+			consumeA(nextPeriodA)
+		case nextPeriodA > nextPeriodB:
+			consumeB(nextPeriodB)
 		default:
-			consumeBoth(nextP)
+			consumeBoth(nextPeriodA)
 		}
 	}
 
 	// consume remaining events in schedule P
-	for iP < lenP {
-		nextP := timeP + periodsP[iP].Length
-		consumeP(nextP)
+	for idxPeriodsA < lenPeriodsA {
+		nextPeriodA := timePeriodsA + periodsA[idxPeriodsA].Length
+		consumeA(nextPeriodA)
 	}
 
 	// consume remaining events in schedule Q
-	for iQ < lenQ {
-		nextQ := timeQ + periodsQ[iQ].Length
-		consumeQ(nextQ)
+	for idxPeriodsB < lenPeriodsB {
+		nextPeriodB := timePeriodsB + periodsB[idxPeriodsB].Length
+		consumeB(nextPeriodB)
 	}
 
 	endTime = time
-	return startTime, endTime, merged
+	return startTime, endTime, conjunctionPeriods
 }
 
-// AlignSchedules rewrites the first period length to align the two arguments to
-// the same start time.
+// AlignSchedules extends the first period's length to align the two given periods
+// to the same start time. The earliest start time is chosen.
+// It returns the aligned new start and end times of the periods.
 func AlignSchedules(
-	startP,
-	startQ int64,
-	p, q []sdkvesting.Period,
+	startTimePeriodA,
+	startTimePeriodB int64,
+	periodsA, periodsB sdkvesting.Periods,
 ) (startTime, endTime int64) {
-	startTime = Min64(startP, startQ)
+	startTime = Min64(startTimePeriodA, startTimePeriodB)
 
-	if len(p) > 0 {
-		p[0].Length += startP - startTime
-	}
-	if len(q) > 0 {
-		q[0].Length += startQ - startTime
+	// add the difference time between
+	if len(periodsA) > 0 {
+		periodsA[0].Length += startTimePeriodA - startTime
 	}
 
-	endP := startTime
-	for _, period := range p {
-		endP += period.Length
+	if len(periodsB) > 0 {
+		periodsB[0].Length += startTimePeriodB - startTime
 	}
-	endQ := startTime
-	for _, period := range q {
-		endQ += period.Length
+
+	endPeriodsA := startTime
+	for _, period := range periodsA {
+		endPeriodsA += period.Length
 	}
-	endTime = Max64(endP, endQ)
-	return
+
+	endPeriodsB := startTime
+	for _, period := range periodsB {
+		endPeriodsB += period.Length
+	}
+
+	// the end time between the 2 periods is the max length
+	endTime = Max64(endPeriodsA, endPeriodsB)
+
+	return startTime, endTime
 }