+++ CHANGELOG.md
@@ -41,6 +41,10 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 - (fees) [\#436](https://github.com/tharsis/evmos/pull/436) Add `x/fees` module.
 
+### Improvements
+
+- (vesting) [\#486](https://github.com/tharsis/evmos/pull/486) Refactor `x/vesting` types and tests.
+
 ### Bug Fixes
 
 - (vesting) [\#483](https://github.com/tharsis/evmos/pull/483) Fix balance clawback when vesting start time is in the future